requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
anyio==4.6.2.post1
async-timeout==4.0.3
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
contourpy==1.3.0
cycler==0.12.1
demjson3==3.0.6
exceptiongroup==1.2.2
fake-useragent==1.5.1
ffmpeg-python==0.2.0
fonttools==4.54.1
frozenlist==1.5.0
future==1.0.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
imageio-ffmpeg==0.5.1
kiwisolver==1.4.7
lxml==5.2.2
matplotlib==3.9.2
multidict==6.1.0
numpy==2.1.3
packaging==24.1
pillow==11.0.0
propcache==0.2.0
py-cord==2.6.1
pycord==0.1.1
pyparsing==3.2.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.2
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
typing_extensions==4.12.2
urllib3==2.2.3
yarl==1.17.1
